requirements.txt
moviepy
opencv-python
pyserial
pyqt5
pillow
numpy
matplotlib
pyaudio
pyttsx3
webrtcvad
vosk[deps]
tritonclient[all]
pyyaml